requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
Flask-PyMongo==2.3.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.0
oauthlib==3.1.0
pandas==1.0.5
pymongo==3.10.1
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
tweepy==3.8.0
urllib3==1.25.9
Werkzeug==1.0.1
